Removal of
H.264 SDP parameters are listed in draft-ietf-rtcweb-video Section 6.2.  RFC 6184 Section 8.2.2 Table 6 indicates how these parameters are to be interpreted for different direction attributes.  Of the parameters listed in draft-ietf-rtcweb-video, only profile-level-id and packetization-mode are listed as a configuration parameter of sent streams.  Therefore only these parameters should be listed in as codec settings.